Feature/#33 bigint

+++ src/main/java/fr/siamois/repositories/VocabularyRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VocabularyRepository extends CrudRepository<Vocabulary, Integer> {
+public interface VocabularyRepository extends CrudRepository<Vocabulary, Long> {
 
 
 }
